requirements.txt
matplotlib==3.3.0
numpy==1.19.1
pandas==1.1.0
pickleshare==0.7.5
Pillow==7.2.0
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
streamlit==0.64.0
urllib3==1.25.10

